+++ Worker/Models/Run.cs
@@ -7,11 +7,9 @@ namespace Worker.Models
         public bool Check { get; set; }
         public bool Inline { get; set; }
         public int Index { get; set; }
-        public int TimeLimit { get; set; }
-        public string Token { get; set; }
 
         public string Stdout { get; set; }
-        public string StdErr { get; set; }
+        public string Stderr { get; set; }
         public Verdict Verdict { get; set; }
         public int? Time { get; set; }
         public int? Memory { set; get; }